updated functions and examples

+++ R/run_fastStructure.R
@@ -21,9 +21,10 @@
 #' @param fastStructure.path Character string with path to folder containing the fastStructure python executable called 'structure.py'
 #' @param cleanup Whether or not the original fastStructure output files (*.log, *.meanQ, *meanP file for each replicate of each K) should be deleted after the data from those files are compiled and saved in three tables. Default TRUE.
 #' @param include.out Character vector indicating which type of files should be included as output. Default is c(".pdf",".Qlog",".margLlog"). An additional file ".Plog" can be included but can be very large.
+#' @param debug Logical indicating whether or not to print messages indicating the internal step of the function.
 #' @return List of plots
 #' @export run_fastStructure
-run_fastStructure <- function(x,format="VCF",coords=NULL,samplenames=NULL,kmax=40,reps=100,save.as=NULL,tolerance=10e-6,prior="simple",full=FALSE,seed=NULL,python.path=NULL,fastStructure.path=NULL,cleanup=TRUE,include.out=c(".pdf",".Qlog",".margLlog")){
+run_fastStructure <- function(x,format="VCF",coords=NULL,samplenames=NULL,kmax=40,reps=100,save.as=NULL,tolerance=10e-6,prior="simple",full=FALSE,seed=NULL,python.path=NULL,fastStructure.path=NULL,cleanup=TRUE,include.out=c(".pdf",".Qlog",".margLlog"), debug=FALSE){
 	if(is.null(save.as)){
 		save.as <- file.path(getwd(),"result_fastStructure.pdf")
 	}
@@ -58,7 +59,7 @@ run_fastStructure <- function(x,format="VCF",coords=NULL,samplenames=NULL,kmax=4
 	}
 	numind      <- length(samplenames)
 	label.size  <- min((288/numind),7)
-		
+	if(debug) message("step 0")
 	### Checking that python2 exists and is executable
 	if(is.null(python.path)){
 		python.testpath <- find.exe.path("python")
@@ -128,6 +129,7 @@ run_fastStructure <- function(x,format="VCF",coords=NULL,samplenames=NULL,kmax=4
 	} else {
 		seed <- NULL
 	}
+	if(debug) message("step 1")
 	outdir.temp <- file.path(dirname(save.as),paste(sample(c(letters,LETTERS,rep(0:9,3)),10,replace=T),collapse=""))
 	dir.create(outdir.temp)
 	for(i in 1:reps){
@@ -141,7 +143,7 @@ run_fastStructure <- function(x,format="VCF",coords=NULL,samplenames=NULL,kmax=4
 		#	run.command2 <- system(command2,intern=TRUE)
 		#}
 	}
-	
+	if(debug) message("step 2")
 	#### Load the marginal likelihood scores (the metric of fit for each K)
 	logpaths <- list.files(outdir.temp,pattern="^.+\\.log$",full.names=TRUE)
 	loglines <- lapply(logpaths,readLines)
@@ -158,17 +160,11 @@ run_fastStructure <- function(x,format="VCF",coords=NULL,samplenames=NULL,kmax=4
 	colnames(margL.mat) <- paste0("rep",1:reps)
 	mean.margL      <- apply(margL.mat,MARGIN=1,FUN=mean,na.rm=TRUE)
 	range.margL.mat <- do.call(rbind,lapply(X=1:nrow(margL.mat),FUN=function(x){range(margL.mat[x,],na.rm=TRUE)}))
-	if(any(diff(mean.margL)<0)){
-		bestK <- unname(which(diff(mean.margL)<0)[1])
-	} else {
-		# bestK <- unname(Krange[1])
-		### might change this back later
-		bestK <- unname(which(mean.margL==max(mean.margL))) 
-	}
+	if(debug) message("step 3")
 	margL.df      <- data.frame(marginalLikelihood=unname(unlist(c(margL.mat))),Kval=rep(Krange,reps))
 	margL.df$Kval <- factor(margL.df$Kval, levels=c(1:nrow(margL.df)))
 	margLPlot     <- ggplot2::ggplot(margL.df, ggplot2::aes(x=Kval, y=marginalLikelihood)) + ggplot2::geom_boxplot(fill='lightgray', outlier.colour="black", outlier.shape=16,outlier.size=2, notch=FALSE) + ggplot2::theme_classic() + ggplot2::labs(title= paste0("marginal likelihood (",reps," replicates) vs. number of ancestral populations (K)"), x="Number of ancestral populations", y = "marginalLikelihood") + ggplot2::theme(plot.title = ggplot2::element_text(hjust = 0.5)) # + ggplot2::geom_vline(xintercept=bestK, linetype=2, color="black", size=0.25)
-	
+	if(debug) message("step 4")
 	#### Load the Q matrices
 	qpaths          <- list.files(outdir.temp,pattern="^.+\\.meanQ$",full.names=TRUE)
 	qmats.list      <- lapply(qpaths,read.table)
@@ -183,10 +179,12 @@ run_fastStructure <- function(x,format="VCF",coords=NULL,samplenames=NULL,kmax=4
 	qmats.list.best <- lapply(1:max(Krange),FUN=function(x){qmats.list[[which(KQmats==x)[bestReps[(x)]]]]})
 	#### List holding population assignment probabilities for each K
 	# Could also try using the mean among qmatrices, but that could be problematic if cluster1 is treated as cluster2 in another iteration.
+	if(debug) message("step 5")
 	slist          <- qmats.list.best
 	Krange.plot    <- setdiff(Krange,1)
 	admixturePlot  <- list(); length(admixturePlot)   <- length(Krange.plot)
 	assignmentPlot <- list(); length(assignmentPlot)  <- length(Krange.plot)
+	if(debug) message("step 6")
 	if(!is.null(coords)){
 		mapplot        <- list(); length(mapplot)     <- length(Krange.plot)
 		x.min <- min((coords[,1]-0.5))
@@ -200,49 +198,65 @@ run_fastStructure <- function(x,format="VCF",coords=NULL,samplenames=NULL,kmax=4
 	} else {
 		mapplot <- NULL
 	}
+	if(debug) message("step 7")
 	for(K in Krange.plot){
+		if(debug) message(paste0("K=",K," step 7.1"))
 		i=(K-1)
-		q.matrix  <- slist[[K]]
-		rownames(q.matrix) <- samplenames
-		colnames(q.matrix) <- paste0("cluster",1:ncol(q.matrix))
-		indv.pop            <- apply(X=q.matrix, MARGIN=1, FUN=function(x){which(x==max(x))})
-		posterior.df       <- data.frame(indv=rep(rownames(q.matrix),ncol(q.matrix)), pop=rep(colnames(q.matrix),each=nrow(q.matrix)), assignment=c(unlist(unname(q.matrix))))
-		posterior.df$indv  <- factor(posterior.df$indv, levels = names(sort(indv.pop)))
 		if(K <= 15){
 			myCols          <- goodcolors2(n=K)
 		}
 		if(K>15){
 			myCols          <- c(goodcolors2(n=15), sample(adegenet::funky(100), size=K-15))
 		}
+		if(debug) message(cat("\r",paste0("K=",K," step 7.2")))
+		q.matrix  <- slist[[K]]
+		rownames(q.matrix) <- samplenames
+		colnames(q.matrix) <- paste0("cluster",1:ncol(q.matrix))
+		indv.pop           <- apply(X=q.matrix, MARGIN=1, FUN=function(x){which(x==max(x))})
+		posterior.df       <- data.frame(indv=rep(rownames(q.matrix),ncol(q.matrix)), pop=rep(colnames(q.matrix),each=nrow(q.matrix)), assignment=c(unlist(unname(q.matrix))))
+		posterior.df$indv  <- factor(posterior.df$indv, levels = names(sort(indv.pop)))
+		if(debug) message(cat("\r",paste0("K=",K," step 7.3")))
 		posterior.gg        <- ggplot2::ggplot(posterior.df, ggplot2::aes(fill= pop, x= assignment, y=indv)) + ggplot2::geom_bar(position="stack", stat="identity") + ggplot2::theme_classic() + ggplot2::theme(axis.text.y = ggplot2::element_text(size = label.size), panel.grid.major = ggplot2::element_blank(), panel.grid.minor = ggplot2::element_blank(), panel.background = ggplot2::element_blank()) + ggplot2::labs(x = "Admixture Proportion",y="",fill="Cluster",title=paste0("K = ",K)) + ggplot2::scale_fill_manual(values=myCols[1:K])
 		admixturePlot[[i]]  <- posterior.gg
+		if(debug) message(cat("\r",paste0("K=",K," step 7.4")))
 		indv.maxPosterior  <- apply(X=q.matrix, MARGIN=1, FUN=function(x){max(x)})
 		labels             <- rep("",nrow(posterior.df))
 		labels[posterior.df[,"assignment"] %in% indv.maxPosterior] <- "+"
 		assignment.K         <- ggplot2::ggplot(data=posterior.df, ggplot2::aes(x= pop, y=indv,fill=assignment)) + ggplot2::geom_tile(color="gray") + ggplot2::theme_classic() + ggplot2::theme(plot.title = ggplot2::element_text(hjust = 0.5), axis.text.y = ggplot2::element_text(size = label.size), panel.grid.major = ggplot2::element_blank(), panel.grid.minor = ggplot2::element_blank(), panel.background = ggplot2::element_blank(), legend.position = "none", ) + ggplot2::labs(title = paste0("K = ",K), x="Clusters", y="") + ggplot2::scale_fill_gradient2(low = "white", mid = "yellow", high = "red", midpoint = 0.5) + ggplot2::geom_text(label=labels)
 		assignmentPlot[[i]]  <- assignment.K
+		if(debug) message(cat("\r",paste0("K=",K," step 7.5")))
 		if(!is.null(coords)){
 			my.palette      <- tess3r::CreatePalette(myCols, 9)
-			mapplot.i       <- tess3r::ggtess3Q(suppressWarnings(tess3r::as.qmatrix(q.matrix)), as.matrix(coords), interpolation.model = tess3r::FieldsKrigModel(10),resolution = c(500,500), col.palette = my.palette, window=c(x.min,x.max,y.min,y.max),background=TRUE, map.polygon=world_sp)
+			if(debug) message(cat("\r",paste0("K=",K," step 7.6")))
+			mapplot.i       <- suppressWarnings(tess3r::ggtess3Q(suppressWarnings(tess3r::as.qmatrix(q.matrix)), as.matrix(coords), interpolation.model = tess3r::FieldsKrigModel(10),resolution = c(500,500), col.palette = my.palette, window=c(x.min,x.max,y.min,y.max),background=TRUE, map.polygon=world_sp))
+			if(debug) message(cat("\r",paste0("K=",K," step 7.7")))
 			mapplot2.i      <- mapplot.i + ggplot2::theme_classic() + ggplot2::labs(title=paste0("Ancestry coefficients; K=",K), x="latitude", y="longitude") + ggplot2::theme(plot.title = ggplot2::element_text(hjust = 0.5), panel.border = ggplot2::element_rect(color = "black", fill=NA, size=1)) + ggplot2::geom_sf(data=world_sf,colour = "black", fill = NA) + ggplot2::coord_sf(xlim=c(x.min, x.max), ylim=c(y.min, ymax=y.max),expand=FALSE)
+			if(debug) message(cat("\r",paste0("K=",K," step 7.8")))
 			mapplot[[i]]    <- mapplot2.i + ggplot2::geom_point(data = coords, ggplot2::aes(x = Lon, y = Lat), size = 1, shape = 21, fill = "black")
 			#mapplot[[i]]    <- mapplot.i + ggplot2::theme_classic() + ggplot2::labs(title=paste0("Ancestry coefficients; K=",K), x="latitude", y="longitude") + ggplot2::theme(plot.title = ggplot2::element_text(hjust = 0.5)) + current.gg.sf + ggplot2::geom_point(data = coords, ggplot2::aes(x = Lon, y = Lat), size = 1, shape = 21, fill = "black")
 		} else {
 			mapplot[[i]] <- NULL
 		}
 	}
+	if(debug) message("step 8")
 	result <- c(list(margLPlot),admixturePlot,assignmentPlot,mapplot)
 	#### Save the PDF
+	if(debug) message("step 9")
 	if(".pdf" %in% include.out){
 	#if(!is.null(save.as)){
 		pdf(height=6,width=10,file=save.as,onefile=TRUE)
 		lapply(X=result,FUN=print)
 		dev.off()
 	}
+	if(debug) message("step 10")
 	if(format=="VCF"){
 		# Delete the temporary fastStructure file
-		remove.str <- file.remove(str.path0)
+		if(file.exists(str.path0)){
+			remove.str <- file.remove(str.path0)
+		}
 	}
+	
+	if(debug) message("step 11")
 	#### Generate compiled output files so that the other output files can be deleted.
 	if(".margLlog" %in% include.out){
 		### Organizing a single data frame to hold all data from all "*.log" files generated by fastStructure. This data frame will be written to a single file to replace the many '*.log' files.
@@ -264,6 +278,7 @@ run_fastStructure <- function(x,format="VCF",coords=NULL,samplenames=NULL,kmax=4
 		}
 		write.table(x=log.df,file=paste0(tools::file_path_sans_ext(save.as),".margLlog"),row.names=F,col.names=T,quote=F,sep="\t")
 	}
+	if(debug) message("step 12")
 	if(".Qlog" %in% include.out){
 		### Organizing a single data frame to hold all q matrices for all replicates of every K. This data frame will be written to a single file so that the many '*.meanQ' files produced by fastStrucure can be deleted.
 		q.df <- NULL
@@ -277,6 +292,7 @@ run_fastStructure <- function(x,format="VCF",coords=NULL,samplenames=NULL,kmax=4
 		}
 		write.table(x=q.df,file=paste0(tools::file_path_sans_ext(save.as),".Qlog"),row.names=F,col.names=T,quote=F,sep="\t")
 	}
+	if(debug) message("step 13")
 	if(".Plog" %in% include.out){
 		### Organizing a single data frame to hold all data from all "*.meanP" files generated by fastStructure. This data frame will be written to a single file to replace the many '*.meanP' files.
 		ppaths <- list.files(outdir.temp,pattern="^.+\\.meanP$",full.names=TRUE)
@@ -300,6 +316,7 @@ run_fastStructure <- function(x,format="VCF",coords=NULL,samplenames=NULL,kmax=4
 		}
 	}
 	# Delete the folder with .log, .meanQ, and .meanP files; the info from these are compiled in a single file.
+	if(debug) message("step 14")
 	if(cleanup){
 		system(paste0("rm -R ",outdir.temp))
 	}
@@ -307,10 +324,16 @@ run_fastStructure <- function(x,format="VCF",coords=NULL,samplenames=NULL,kmax=4
 }
 #' @examples
 #' library(misc.wrappers)
-#' # Define path to input VCF file containing similated data for 500 SNPs from 50 individuals in three populations.
-#' example_vcf_path <- file.path(system.file("extdata", package = "misc.wrappers"),"simulated_K4.vcf.gz")
-#' # Perform sNMF analyses on the simulated dataset for for K=2–10 and 30 replicates.
-#' test_fastStructure_K4 <- run_fastStructure(x=example_vcf_path, kmax=10, reps=30, save.as="fastStructure_example_K4.pdf",include.out=c(".pdf"))
+#' ## Example 1:
+#' # Path to VCF with SNPs
+#' vcf.path    <- file.path(system.file("extdata", package = "misc.wrappers"),"simK4.vcf.gz")
+#' # Run fastStructure 30 times each for K=1-10
+#' run_fastStructure(x=vcf.path,kmax=10,reps=30,save.as="fs_simK4.pdf",include.out=c(".pdf"))
+#' 
+#' ## Example 2: Same SNP dataset as example 1, but here we also provide Lon/Lat coordinates of individuals to geographically interpolate admixture coefficients.
+#' vcf.path    <- file.path(system.file("extdata", package = "misc.wrappers"), "simK4.vcf.gz")
+#' coords.path <- file.path(system.file("extdata", package = "misc.wrappers"), "simK4_coords.txt")
+#' run_fastStructure(x=vcf.path, coords=coords.path, kmax=10, reps=30, save.as="fs_simK4_withCoords.pdf", include.out=c(".pdf"))
 
 
 
@@ -556,7 +579,7 @@ vcf2structure <- function(vcf, IndvNames=TRUE, OneRowPerIndv=TRUE, MarkerNames=T
 }
 #' @examples
 #' library(misc.wrappers)
-#' vcf2structure(vcf="Ahaetulla-prasina_AllPops_BestSNP.vcf",out="example.str")
+
 
 
 #' @title Convert VCF file/object to fastStructure input file (for SNP data)
